Refactor helper module and update unit tests
